Created the grader class
This class has two arrays, one for scores and another for storing the indecies of the top three highest scored roommate answers.
The score for the answers is done by comparing strings, incrementing by one when a character matches in a string.
The indicies stored in the matchIndicies is used to find the first and last names in the roommate.csv file